Refactored core module into models and handler
